# internal/domain/client.go
package domain

import "time"

type Client struct {
	ID             string     `db:"id"`
	Fingerprint    string     `db:"fingerprint"`
	HostName       string     `db:"hostname"`
	SessionToken   string     `db:"session_token"`
	TokenExpiresAt *time.Time `db:"token_expires_at"`
	CreatedAt      time.Time  `db:"created_at"`
	LastSeenAt     *time.Time `db:"last_seen_at"`
	OS             string     `db:"os"`
}

type ClientSummary struct {
	ID         string     `json:"id"`
	HostName   string     `json:"hostname"`
	CreatedAt  time.Time  `json:"created_at"`
	LastSeenAt *time.Time `json:"last_seen_at"`
	Online     bool       `json:"online"`
	OS         string     `json:"operating_system"`
}

type ClientSession struct {
	ID             string     `db:"id"`
	ClientID       string     `db:"client_id"`
	ConnectedAt    time.Time  `db:"connected_at"`
	DisconnectedAt *time.Time `db:"disconnected_at"`
}